fixing  `float32` vs `float64` assert error

+++ test/test_flowgraph.py
@@ -9,20 +9,25 @@ class Test_flowgraph:
 
     def test_one(self):
         x = [5,5]
-        assert self.model.predict([x])[0][0].round(5) == self.flowgraph.propagate([x])[0][0].round(5)
+        assert (self.model.predict([x])[0][0].astype('float32').round(5) 
+                == self.flowgraph.propagate([x])[0][0].astype('float32').round(5))
 
     def test_two(self):
         x = [1,2]
-        assert self.model.predict([x])[0][0].round(5) == self.flowgraph.propagate([x])[0][0].round(5)
+        assert (self.model.predict([x])[0][0].astype('float32').round(5) 
+                == self.flowgraph.propagate([x])[0][0].astype('float32').round(5))
         
     def test_three(self):
         x = [10,0]
-        assert self.model.predict([x])[0][0].round(5) == self.flowgraph.propagate([x])[0][0].round(5)
+        assert (self.model.predict([x])[0][0].astype('float32').round(5) 
+                == self.flowgraph.propagate([x])[0][0].astype('float32').round(5))
     
     def test_four(self):
         x = [3,2]
-        assert self.model.predict([x])[0][0].round(5) == self.flowgraph.propagate([x])[0][0].round(5)
+        assert (self.model.predict([x])[0][0].astype('float32').round(5) 
+                == self.flowgraph.propagate([x])[0][0].astype('float32').round(5))
 
     def test_five(self):
         x = [8,2]
-        assert self.model.predict([x])[0][0].round(5) == self.flowgraph.propagate([x])[0][0].round(5)
+        assert (self.model.predict([x])[0][0].astype('float32').round(5) 
+                == self.flowgraph.propagate([x])[0][0].astype('float32').round(5))
